+++ plugins/main.lua
@@ -15,6 +15,7 @@ Classes = {
     "Class",
     "Color",
     "Damager",
+    "Difficulty",
     "Equipment",
     "Helper",
     "Initialize",
@@ -57,6 +58,7 @@ function __initialize()
     Actor.__initialize()
     Artifact.__initialize()
     Buff.__initialize()
+    Difficulty.__initialize()
     Equipment.__initialize()
     Instance.__initialize()
     Interactable.__initialize()